+++ examples-server/hono-s3/src/db.ts
@@ -10,41 +10,33 @@ const __dirname = path.dirname(fileURLToPath(import.meta.url));
 // Load optional .env.hotupdater file for local development
 config({ path: path.join(__dirname, ".env.hotupdater") });
 
-const region = process.env.AWS_REGION || "us-east-1";
-const endpoint = process.env.AWS_S3_ENDPOINT || "http://localhost:4566";
-const accessKeyId = process.env.AWS_ACCESS_KEY_ID || "test";
-const secretAccessKey = process.env.AWS_SECRET_ACCESS_KEY || "test";
-const metadataBucket =
-  process.env.AWS_S3_METADATA_BUCKET || "hot-updater-metadata";
-const bundlesBucket =
-  process.env.AWS_S3_BUNDLES_BUCKET || "hot-updater-bundles";
+const options =
+  process.env.NODE_ENV === "test"
+    ? {
+        region: process.env.AWS_REGION || "us-east-1",
+        endpoint: process.env.AWS_S3_ENDPOINT || "http://localhost:4566",
+        credentials: {
+          accessKeyId: process.env.AWS_ACCESS_KEY_ID || "test",
+          secretAccessKey: process.env.AWS_SECRET_ACCESS_KEY || "test",
+        },
+        bucketName:
+          process.env.AWS_S3_METADATA_BUCKET || "hot-updater-metadata",
+        // localstack s3
+        forcePathStyle: true,
+      }
+    : {
+        region: "auto",
+        endpoint: process.env.R2_ENDPOINT,
+        credentials: {
+          accessKeyId: process.env.R2_ACCESS_KEY_ID!,
+          secretAccessKey: process.env.R2_SECRET_ACCESS_KEY!,
+        },
+        bucketName: process.env.R2_BUCKET_NAME!,
+      };
 
 export const hotUpdater = createHotUpdater({
-  database: s3Database({
-    region,
-    endpoint,
-    credentials: {
-      accessKeyId,
-      secretAccessKey,
-    },
-    bucketName: metadataBucket,
-    // localstack s3
-    forcePathStyle: true,
-  }),
-  storages: [
-    mockStorage({}),
-    s3Storage({
-      region,
-      endpoint,
-      credentials: {
-        accessKeyId,
-        secretAccessKey,
-      },
-      bucketName: bundlesBucket,
-      // localstack s3
-      forcePathStyle: true,
-    }),
-  ],
+  database: s3Database(options),
+  storages: [mockStorage({}), s3Storage(options)],
   basePath: "/hot-updater",
 });