+++ packages/cli-tools/CHANGELOG.md
@@ -1,5 +1,33 @@
 # @hot-updater/cli-tools
 
+## 0.24.7
+
+### Patch Changes
+
+- 294e324: fix: update babel plugin path in documentation and plugin files
+- Updated dependencies [294e324]
+  - @hot-updater/plugin-core@0.24.7
+
+## 0.24.6
+
+### Patch Changes
+
+- 9d7b6af: feat(aws): sso template with fromSSO
+  - @hot-updater/plugin-core@0.24.6
+
+## 0.24.5
+
+### Patch Changes
+
+- @hot-updater/plugin-core@0.24.5
+
+## 0.24.4
+
+### Patch Changes
+
+- Updated dependencies [7ed539f]
+  - @hot-updater/plugin-core@0.24.4
+
 ## 0.24.3
 
 ### Patch Changes