+++ packages/hot-updater/CHANGELOG.md
@@ -1,5 +1,52 @@
 # hot-updater
 
+## 0.24.7
+
+### Patch Changes
+
+- 294e324: fix: update babel plugin path in documentation and plugin files
+- Updated dependencies [294e324]
+  - @hot-updater/cli-tools@0.24.7
+  - @hot-updater/console@0.24.7
+  - @hot-updater/core@0.24.7
+  - @hot-updater/server@0.24.7
+  - @hot-updater/plugin-core@0.24.7
+
+## 0.24.6
+
+### Patch Changes
+
+- 9d7b6af: feat(aws): sso template with fromSSO
+- 962ecdd: fix(expo): fingerprint autolinking for expo
+- Updated dependencies [9d7b6af]
+  - @hot-updater/cli-tools@0.24.6
+  - @hot-updater/console@0.24.6
+  - @hot-updater/server@0.24.6
+  - @hot-updater/core@0.24.6
+  - @hot-updater/plugin-core@0.24.6
+
+## 0.24.5
+
+### Patch Changes
+
+- f755c3c: Add build\* to default fingerprint ignore paths
+  - @hot-updater/cli-tools@0.24.5
+  - @hot-updater/console@0.24.5
+  - @hot-updater/core@0.24.5
+  - @hot-updater/server@0.24.5
+  - @hot-updater/plugin-core@0.24.5
+
+## 0.24.4
+
+### Patch Changes
+
+- Updated dependencies [7ed539f]
+  - @hot-updater/plugin-core@0.24.4
+  - @hot-updater/cli-tools@0.24.4
+  - @hot-updater/console@0.24.4
+  - @hot-updater/server@0.24.4
+  - @hot-updater/core@0.24.4
+
 ## 0.24.3
 
 ### Patch Changes